Refine MailcapCommandMap class loading tests
