+++ tests/src/javax.activation/activation/1.1/src/test/java/javax_activation/activation/MailcapCommandMapTest.java
@@ -6,7 +6,13 @@
  */
 package javax_activation.activation;
 
+import java.awt.datatransfer.DataFlavor;
+import java.awt.datatransfer.UnsupportedFlavorException;
+import java.io.IOException;
+import java.io.OutputStream;
+
 import javax.activation.DataContentHandler;
+import javax.activation.DataSource;
 import javax.activation.MailcapCommandMap;
 
 import org.junit.jupiter.api.Test;
@@ -15,12 +21,33 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 public class MailcapCommandMapTest {
     @Test
+    void createDataContentHandlerLoadsHandlerFromContextClassLoader() {
+        ClassLoader originalClassLoader = Thread.currentThread().getContextClassLoader();
+        ClassLoader loadingClassLoader = new ClassLoader(MailcapCommandMapTest.class.getClassLoader()) {
+        };
+
+        Thread.currentThread().setContextClassLoader(loadingClassLoader);
+        try {
+            MailcapCommandMap commandMap = new MailcapCommandMap();
+            commandMap.addMailcap(
+                "application/x-mailcap-test-load;; x-java-content-handler=" + SampleDataContentHandler.class.getName()
+            );
+
+            DataContentHandler handler = commandMap.createDataContentHandler("application/x-mailcap-test-load");
+
+            assertThat(handler).isInstanceOf(SampleDataContentHandler.class);
+        } finally {
+            Thread.currentThread().setContextClassLoader(originalClassLoader);
+        }
+    }
+
+    @Test
     void createDataContentHandlerFallsBackToClassForName() {
         ClassLoader originalClassLoader = Thread.currentThread().getContextClassLoader();
-        ClassLoader blockingClassLoader = new ClassLoader(originalClassLoader) {
+        ClassLoader blockingClassLoader = new ClassLoader(MailcapCommandMapTest.class.getClassLoader()) {
             @Override
             public Class<?> loadClass(String name) throws ClassNotFoundException {
-                if (MailcapCommandMapDataContentHandler.class.getName().equals(name)) {
+                if (SampleDataContentHandler.class.getName().equals(name)) {
                     throw new ClassNotFoundException(name);
                 }
                 return super.loadClass(name);
@@ -31,15 +58,40 @@ public class MailcapCommandMapTest {
         try {
             MailcapCommandMap commandMap = new MailcapCommandMap();
             commandMap.addMailcap(
-                "application/x-mailcap-test;; x-java-content-handler=" + MailcapCommandMapDataContentHandler.class.getName()
+                "application/x-mailcap-test-fallback;; x-java-content-handler=" + SampleDataContentHandler.class.getName()
             );
 
-            DataContentHandler handler = commandMap.createDataContentHandler("application/x-mailcap-test");
+            DataContentHandler handler = commandMap.createDataContentHandler("application/x-mailcap-test-fallback");
 
-            assertThat(handler).isInstanceOf(MailcapCommandMapDataContentHandler.class);
+            assertThat(handler).isInstanceOf(SampleDataContentHandler.class);
         } finally {
             Thread.currentThread().setContextClassLoader(originalClassLoader);
         }
     }
 
+    public static final class SampleDataContentHandler implements DataContentHandler {
+        public SampleDataContentHandler() {
+        }
+
+        @Override
+        public DataFlavor[] getTransferDataFlavors() {
+            return new DataFlavor[0];
+        }
+
+        @Override
+        public Object getTransferData(DataFlavor flavor, DataSource dataSource)
+                throws UnsupportedFlavorException, IOException {
+            throw new UnsupportedFlavorException(flavor);
+        }
+
+        @Override
+        public Object getContent(DataSource dataSource) {
+            return null;
+        }
+
+        @Override
+        public void writeTo(Object object, String mimeType, OutputStream outputStream) throws IOException {
+            outputStream.write(new byte[0]);
+        }
+    }
 }
